refactor: move generate PR body logic to separate file
